requirements.txt
# Backend — FastAPI
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0

# Auth
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# Database
sqlmodel>=0.0.16
sqlalchemy>=2.0

# LLM Providers
openai>=1.30
anthropic>=0.25
google-generativeai>=0.5

# PII Detection
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
spacy>=3.7

# Token Counting & Cost
tiktoken>=0.7

# Encryption
cryptography>=42.0

# Logging
python-json-logger>=2.0

# Frontend (legacy — Streamlit)
streamlit>=1.20
requests>=2.28

# Testing
pytest>=8.0
httpx>=0.27
